backend/requirements.txt
flask>=2.0.0
google-cloud-firestore
firebase-admin
